Revocation logic for authentication
Adds a ?revoked parameter to validation functions, and a ?crls to chain_of_trust
authenticator.  CRL.is_revoked is a convenience function which gets a list of
revocation lists and an issuer and a certificate, and returns true if the
certificate is revoked, false otherwise.

Furthermore, this commit adds some tests for revocation which also use the
builtin CA signing mechanisms to create some certificates on the fly.  This
could be extended further (e.g. using crowbar or other randomised testing).