Rename Authenticator.server_{cert,key}_fingerprint
By dropping the server_ prefix. The functions do not check that the
certificate have extended key usage of server auth and could just as
well be used to authenticate clients (although the server case is more
likely). The old names are kept as aliases and a deprecation attribute
is added.